代码提交

+++ Src/BudgetSystem/BudgetSystem/Budget/frmBudgetQuery.cs
@@ -43,6 +43,7 @@ namespace BudgetSystem
             base.InitModelOperate();
             this.ModelOperateRegistry.Add(ModelOperateHelper.GetOperate(OperateTypes.New));
             this.ModelOperateRegistry.Add(ModelOperateHelper.GetOperate(OperateTypes.Modify));
+            this.ModelOperateRegistry.Add(ModelOperateHelper.GetOperate(OperateTypes.ModifyEx, "修改其他供应商"));
             this.ModelOperateRegistry.Add(ModelOperateHelper.GetOperate(OperateTypes.SubmitApply, "提交流程"));
             this.ModelOperateRegistry.Add(ModelOperateHelper.GetOperate(OperateTypes.ModifyApply, "申请修改"));
             this.ModelOperateRegistry.Add(ModelOperateHelper.GetOperate(OperateTypes.DeleteApply, "申请删除"));
@@ -87,6 +88,10 @@ namespace BudgetSystem
             else if (operate.Operate == OperateTypes.Modify.ToString())
             {
                 ModifyBudget();
+            }
+            else if (operate.Operate == OperateTypes.ModifyEx.ToString())
+            {
+                ModifyBudgetSlipper();
             }
             else if (operate.Operate == OperateTypes.SubmitApply.ToString())
             {
@@ -211,7 +216,11 @@ namespace BudgetSystem
 
         private void ShowBudgetAccountBillView()
         {
-            Budget budget = this.gvBudget.GetFocusedRow() as Budget;
+            if (this.gvBudget.FocusedRowHandle < 0)
+            {
+                return;
+            }
+            Budget budget = this.gvBudget.GetRow(this.gvBudget.FocusedRowHandle) as Budget;
             if (budget != null)
             {
                 frmAccountBillView form = new frmAccountBillView();
@@ -240,46 +249,99 @@ namespace BudgetSystem
 
         private void ModifyBudget()
         {
-            Budget budget = this.gvBudget.GetFocusedRow() as Budget;
-            if (budget != null)
-            {
-                string message = string.Empty;
-                if (budget.EnumFlowState == EnumDataFlowState.审批中
-                    || (EnumFlowNames.预算单审批流程.ToString().Equals(budget.FlowName) && budget.EnumFlowState == EnumDataFlowState.审批通过))
-                {
-                    message = string.Format("{0}的预算单不能修改。", budget.EnumFlowState.ToString());
-                }
-                else if (EnumFlowNames.预算单修改流程.ToString().Equals(budget.FlowName) && budget.EnumFlowState != EnumDataFlowState.审批通过)
-                {
-                    message = string.Format("{0}未审批通过不能修改。", EnumFlowNames.预算单修改流程.ToString());
-                }
-                if (!RunInfo.Instance.CurrentUser.UserName.Equals(budget.Salesman))
-                {
-                    message = string.Format("当前为{0}的预算单，{1}不能修改。", RunInfo.Instance.CurrentUser.RealName, budget.SalesmanName);
-                }
-                if (!string.IsNullOrEmpty(message))
-                {
-                    XtraMessageBox.Show(message, "提示");
-                    this.ViewBudget();
-                    return;
-                }
-                frmBudgetEdit form = new frmBudgetEdit();
-                form.WorkModel = EditFormWorkModels.Modify;
-                form.CurrentBudget = budget;
-                if (form.ShowDialog(this) == System.Windows.Forms.DialogResult.OK)
-                {
-                    this.RefreshData();
-                }
-            }
-            else
+            if (this.gvBudget.FocusedRowHandle < 0)
             {
                 XtraMessageBox.Show("请选择需要修改的项");
+                return;
+            }
+            Budget budget = this.gvBudget.GetRow(this.gvBudget.FocusedRowHandle) as Budget;
+            if (budget == null)
+            {
+                XtraMessageBox.Show("请选择需要修改的项");
+                return;
+            }
+            budget = bm.GetBudget(budget.ID);
+            if (budget == null)
+            {
+                XtraMessageBox.Show("选择修改的项已经不存在。");
+                return;
+            }
+            string message = string.Empty;
+            if (budget.EnumFlowState == EnumDataFlowState.审批中
+                || (EnumFlowNames.预算单审批流程.ToString().Equals(budget.FlowName) && budget.EnumFlowState == EnumDataFlowState.审批通过))
+            {
+                message = string.Format("{0}的预算单不能修改。", budget.EnumFlowState.ToString());
+            }
+            else if (EnumFlowNames.预算单修改流程.ToString().Equals(budget.FlowName) && budget.EnumFlowState != EnumDataFlowState.审批通过)
+            {
+                message = string.Format("{0}未审批通过不能修改。", EnumFlowNames.预算单修改流程.ToString());
+            }
+            if (!RunInfo.Instance.CurrentUser.UserName.Equals(budget.Salesman))
+            {
+                message = string.Format("当前为{0}的预算单，{1}不能修改。", RunInfo.Instance.CurrentUser.RealName, budget.SalesmanName);
+            }
+            if (!string.IsNullOrEmpty(message))
+            {
+                XtraMessageBox.Show(message, "提示");
+                this.ViewBudget();
+                return;
+            }
+            frmBudgetEdit form = new frmBudgetEdit();
+            form.WorkModel = EditFormWorkModels.Modify;
+            form.CurrentBudget = budget;
+            if (form.ShowDialog(this) == System.Windows.Forms.DialogResult.OK)
+            {
+                this.RefreshData();
+            }
+        }
+
+        private void ModifyBudgetSlipper()
+        {
+            if (this.gvBudget.FocusedRowHandle < 0)
+            {
+                XtraMessageBox.Show("请选择需要修改的项。");
+                return;
+            }
+            Budget budget = this.gvBudget.GetRow(this.gvBudget.FocusedRowHandle) as Budget;
+            if (budget == null)
+            {
+                XtraMessageBox.Show("请选择需要修改的项。");
+                return;
+            }
+            budget = bm.GetBudget(budget.ID);
+            if (budget == null)
+            {
+                XtraMessageBox.Show("请选择修改的项已经不存在。");
+                return;
+            }
+            string message = string.Empty;
+            if (!RunInfo.Instance.CurrentUser.UserName.Equals(budget.Salesman))
+            {
+                message = string.Format("当前为{0}的预算单，{1}不能修改。", RunInfo.Instance.CurrentUser.RealName, budget.SalesmanName);
+            }
+            if (!string.IsNullOrEmpty(message))
+            {
+                XtraMessageBox.Show(message, "提示");
+                this.ViewBudget();
+                return;
+            }
+            frmBudgetEdit form = new frmBudgetEdit();
+            form.WorkModel = EditFormWorkModels.Custom;
+            form.CurrentBudget = budget;
+            if (form.ShowDialog(this) == System.Windows.Forms.DialogResult.OK)
+            {
+                this.RefreshData();
             }
         }
 
         private void SubmitApply()
         {
-            Budget budget = this.gvBudget.GetFocusedRow() as Budget;
+            if (this.gvBudget.FocusedRowHandle < 0)
+            {
+                XtraMessageBox.Show("请选择需要提交流程的项");
+                return;
+            }
+            Budget budget = this.gvBudget.GetRow(this.gvBudget.FocusedRowHandle) as Budget;
             if (budget != null)
             {
                 if (budget.EnumFlowState == EnumDataFlowState.审批中)
@@ -306,7 +368,12 @@ namespace BudgetSystem
 
         private void ModifyApply()
         {
-            Budget budget = this.gvBudget.GetFocusedRow() as Budget;
+            if (this.gvBudget.FocusedRowHandle < 0)
+            {
+                XtraMessageBox.Show("请选择需要申请修改的项");
+                return;
+            }
+            Budget budget = this.gvBudget.GetRow(this.gvBudget.FocusedRowHandle) as Budget;
             if (budget != null)
             {
                 budget = bm.GetBudget(budget.ID);
@@ -350,7 +417,12 @@ namespace BudgetSystem
 
         private void DeleteApply()
         {
-            Budget budget = this.gvBudget.GetFocusedRow() as Budget;
+            if (this.gvBudget.FocusedRowHandle < 0)
+            {
+                XtraMessageBox.Show("请选择需要申请删除的项");
+                return;
+            }
+            Budget budget = this.gvBudget.GetRow(this.gvBudget.FocusedRowHandle) as Budget;
             if (budget != null)
             {
                 if (budget.EnumFlowState == EnumDataFlowState.审批中)
@@ -371,6 +443,7 @@ namespace BudgetSystem
                     {
                         XtraMessageBox.Show("提交流程成功。");
                         LoadData();
+                        XtraMessageBox.Show("请选择需要申请删除的项");
                     }
                     else
                     {
@@ -386,7 +459,12 @@ namespace BudgetSystem
 
         private void ViewBudget()
         {
-            Budget budget = this.gvBudget.GetFocusedRow() as Budget;
+            if (this.gvBudget.FocusedRowHandle < 0)
+            {
+                XtraMessageBox.Show("请选择需要查看详情的项");
+                return;
+            }
+            Budget budget = this.gvBudget.GetRow(this.gvBudget.FocusedRowHandle) as Budget;
             if (budget == null)
             {
                 XtraMessageBox.Show("请选择需要查看详情的项");
@@ -406,7 +484,12 @@ namespace BudgetSystem
 
         private void ViewApplyHistory()
         {
-            Budget budget = this.gvBudget.GetFocusedRow() as Budget;
+            if (this.gvBudget.FocusedRowHandle < 0)
+            {
+                XtraMessageBox.Show("请选择需要查看审批状态的项");
+                return;
+            }
+            Budget budget = this.gvBudget.GetRow(this.gvBudget.FocusedRowHandle) as Budget;
             if (budget != null)
             {
                 frmHistory hisotryForm = new frmHistory();
@@ -421,7 +504,12 @@ namespace BudgetSystem
 
         private void FinalAccount()
         {
-            Budget budget = this.gvBudget.GetFocusedRow() as Budget;
+            if (this.gvBudget.FocusedRowHandle < 0)
+            {
+                XtraMessageBox.Show("请选择需要查看单一合同决算报表的项");
+                return;
+            }
+            Budget budget = this.gvBudget.GetRow(this.gvBudget.FocusedRowHandle) as Budget;
             if (budget != null)
             {
                 frmSingleBudgetFinalAccountsReport finalAccountReportForm = new frmSingleBudgetFinalAccountsReport();
@@ -436,17 +524,25 @@ namespace BudgetSystem
 
         private void DeleteBudget()
         {
-            Budget budget = this.gvBudget.GetFocusedRow() as Budget;
+            if (this.gvBudget.FocusedRowHandle < 0)
+            {
+                XtraMessageBox.Show("请选择需要删除的项");
+                return;
+            }
+            Budget budget = this.gvBudget.GetRow(this.gvBudget.FocusedRowHandle) as Budget;
             if (budget != null)
             {
-                string message = this.bm.DeleteBudget(budget.ID);
-                if (!string.IsNullOrEmpty(message))
+                if (XtraMessageBox.Show(string.Format("是否真的要删除【{0}】合同？删除后将无法恢复。", budget.ContractNO), "提示", MessageBoxButtons.YesNo) == System.Windows.Forms.DialogResult.Yes)
                 {
-                    XtraMessageBox.Show(message, "提示");
-                }
-                else
-                {
-                    this.RefreshData();
+                    string message = this.bm.DeleteBudget(budget.ID);
+                    if (!string.IsNullOrEmpty(message))
+                    {
+                        XtraMessageBox.Show(message, "提示");
+                    }
+                    else
+                    {
+                        this.RefreshData();
+                    }
                 }
             }
             else
@@ -473,13 +569,18 @@ namespace BudgetSystem
 
         protected override void InitGridViewAction()
         {
-            this.gridViewAction.Add(this.gvBudget, new ActionWithPermission() { MainAction = ModifyBudget, MainOperate = OperateTypes.Modify, SecondAction = ViewBudget, SecondOperate = OperateTypes.View });
+            this.gridViewAction.Add(this.gvBudget, new ActionWithPermission() { MainAction = ModifyBudget, MainOperate = OperateTypes.Modify, SecondAction = ModifyBudgetSlipper, SecondOperate = OperateTypes.ModifyEx });
 
         }
 
         protected override void PrintItem()
         {
-            Budget budget = this.gvBudget.GetFocusedRow() as Budget;
+            if (this.gvBudget.FocusedRowHandle < 0)
+            {
+                XtraMessageBox.Show("请选择需要打印的项");
+                return;
+            }
+            Budget budget = this.gvBudget.GetRow(this.gvBudget.FocusedRowHandle) as Budget;
             if (budget != null)
             {
                 frmBudgetEdit form = new frmBudgetEdit();
@@ -499,7 +600,12 @@ namespace BudgetSystem
         /// </summary>
         private void RejectedAccount()
         {
-            Budget budget = this.gvBudget.GetFocusedRow() as Budget;
+            if (this.gvBudget.FocusedRowHandle < 0)
+            {
+                XtraMessageBox.Show("请选择需要驳回结账申请的项");
+                return;
+            }
+            Budget budget = this.gvBudget.GetRow(this.gvBudget.FocusedRowHandle) as Budget;
             if (budget != null)
             {
                 if (budget.EnumState != EnumBudgetState.财务归档征求)
@@ -550,7 +656,12 @@ namespace BudgetSystem
             }
 
 
-            //Budget budget = this.gvBudget.GetFocusedRow() as Budget;
+            //if (this.gvBudget.FocusedRowHandle < 0)
+            //{
+            //    XtraMessageBox.Show("请选择需要财务归档征求的项");
+            //    return;
+            //}
+            //Budget budget = this.gvBudget.GetRow(this.gvBudget.FocusedRowHandle) as Budget; ;
             //if (budget != null)
             //{
             //    if (budget.EnumState != EnumBudgetState.进行中)
@@ -580,7 +691,12 @@ namespace BudgetSystem
         /// </summary>
         private void Archive()
         {
-            Budget budget = this.gvBudget.GetFocusedRow() as Budget;
+            if (this.gvBudget.FocusedRowHandle < 0)
+            {
+                XtraMessageBox.Show("请选择需要归档的项");
+                return;
+            }
+            Budget budget = this.gvBudget.GetRow(this.gvBudget.FocusedRowHandle) as Budget;
             if (budget != null)
             {
                 budget = bm.GetBudget(budget.ID);
@@ -613,7 +729,12 @@ namespace BudgetSystem
 
         private void AddPaymentNote()
         {
-            Budget budget = this.gvBudget.GetFocusedRow() as Budget;
+            if (this.gvBudget.FocusedRowHandle < 0)
+            {
+                XtraMessageBox.Show("请选择合同");
+                return;
+            }
+            Budget budget = this.gvBudget.GetRow(this.gvBudget.FocusedRowHandle) as Budget;
             if (budget == null)
             {
                 XtraMessageBox.Show("请选择合同");
